pyproject.toml
[project]
name = "axis-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.20.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "websockets>=16.0",
]
